Exercise 10.23: sorting the reviewed repositories list
Latest repositories. The repository with the latest first review is on the top of the list. (default)
Highest rated repositories. The repository with the highest average rating is on the top of the list.
Lowest rated repositories. The repository with the lowest average rating is on the top of the list.