Exercise 13.19.
Give users the ability to add blogs on the system to a reading list.
When added to the reading list, the blog should be in the unread state.
The blog can later be marked as read. Implement the reading list using a
connection table. Make database changes using migrations.

In this task, adding to a reading list and displaying the list need not
be successful other than directly using the database.