Exercise 13.20.
Add functionality to the application to support the reading list.

Adding a blog to the reading list is done by making an HTTP POST to the
path /api/readinglists, the request will be accompanied with the blog
and user id:

{
  "blogId": 10,
  "userId": 3
}

Also modify the individual user route GET /api/users/:id to return not
only the user's other information but also the reading list, e.g. in the
following format:

{
  name: "Matti Luukkainen",
  username: "mluukkai@iki.fi",
  readings: [
    {
      id: 3,
      url: "https://google.com",
      title: "Clean React",
      author: "Dan Abramov",
      likes: 34,
      year: null,
    },
    {
      id: 4,
      url: "https://google.com",
      title: "Clean Code",
      author: "Bob Martin",
      likes: 5,
      year: null,
    }
  ]
}