13.22 - 13.23
Exercise 13.22.

Implement functionality in the application to mark a blog in the reading
list as read. Marking as read is done by making a request to the PUT
/api/readinglists/:id path, and sending the request with
{ "read": true }

The user can only mark the blogs in their own reading list as read. The
user is identified as usual from the token accompanying the request.

Exercise 13.23.

Modify the route that returns a single user's information so that the
request can control which of the blogs in the reading list are returned:

GET /api/users/:id returns the entire reading list
GET /api/users/:id?read=true returns blogs that have been read
GET /api/users/:id?read=false returns blogs that have not been read