Exercise 13.24.
Now expand the system so that the user who has lost access will not be
able to perform any actions that require login.

You will probably need at least the following for the implementation

a boolean value column in the user table to indicate whether the user is
disabled

it is sufficient to disable and enable users directly from the database
a table that stores active sessions

a session is stored in the table when a user logs in, i.e. operation
POST /api/login
the existence (and validity) of the session is always checked when the
user makes an operation that requires login
a route that allows the user to "log out" of the system, i.e. to
practically remove active sessions from the database, the route can be
e.g. DELETE /api/logout
Keep in mind that actions requiring login should not be successful with
an "expired token", i.e. with the same token after logging out.